cart/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
redis==5.0.1
gunicorn==21.2.0
openfeature-sdk==0.8.4
openfeature-provider-flagd==0.2.6

